ght2dm: improve performances
When using ght2dm on a huge amount of data, insertion become really slow
into the PostgreSQL database. This commit address this problem by
inserting all the documents of a BSON dump into a single transaction.
Update operation has been removed since this is unlikely that the data
in the BSON dump are more up to date that the ones already in the
database. Some of the constraints are disabled too, which should improve
performances as well.